requirements.txt
pillow
fpdf2
natsort
py7zr
PyPDF2
PyQt5
